+++ tests/test_cloudformation/test_cloudformation_stack_crud.py
@@ -21,7 +21,6 @@ from moto import (
     mock_route53_deprecated,
     mock_iam_deprecated,
     mock_dynamodb2,
-    mock_cloudformation,
 )
 from moto.cloudformation import cloudformation_backends
 
@@ -73,6 +72,7 @@ dummy_template4 = {
 dummy_template_json = json.dumps(dummy_template)
 dummy_template_json2 = json.dumps(dummy_template2)
 dummy_template_json3 = json.dumps(dummy_template3)
+dummy_template_json4 = json.dumps(dummy_template4)
 
 
 @mock_cloudformation_deprecated
@@ -218,27 +218,13 @@ def test_describe_stack_by_stack_id():
 def test_delete_stack_dynamo_template():
     conn = boto.connect_cloudformation()
     dynamodb_client = boto3.client("dynamodb", region_name="us-east-1")
-    conn.create_stack("test_stack", template_body=dummy_template4)
+    conn.create_stack("test_stack", template_body=dummy_template_json4)
     table_desc = dynamodb_client.list_tables()
     len(table_desc.get("TableNames")).should.equal(1)
     conn.delete_stack("test_stack")
     table_desc = dynamodb_client.list_tables()
     len(table_desc.get("TableNames")).should.equal(0)
-    conn.create_stack("test_stack", template_body=dummy_template4)
-
-
-@mock_dynamodb2
-@mock_cloudformation
-def test_delete_stack_dynamo_template():
-    conn = boto3.client("cloudformation", region_name="us-east-1")
-    dynamodb_client = boto3.client("dynamodb", region_name="us-east-1")
-    conn.create_stack(StackName="test_stack", TemplateBody=json.dumps(dummy_template4))
-    table_desc = dynamodb_client.list_tables()
-    len(table_desc.get("TableNames")).should.equal(1)
-    conn.delete_stack(StackName="test_stack")
-    table_desc = dynamodb_client.list_tables()
-    len(table_desc.get("TableNames")).should.equal(0)
-    conn.create_stack(StackName="test_stack", TemplateBody=json.dumps(dummy_template4))
+    conn.create_stack("test_stack", template_body=dummy_template_json4)
 
 
 @mock_cloudformation_deprecated
